Refactoring contracts logic into separate modules
